added edit modal component
created a modal that is an overlay of the whole page and it will allow users to edit their account details
