+++ src/components/ProgressBar.js
@@ -5,6 +5,7 @@ import { motion } from 'framer-motion'
 const ProgressBar = ({ progress }) => {
     return (
         <div className="progressContainer">
+            <p>{progress !== 0 && progress.toFixed(0) + '%'}</p>
             <div className="progress" style={{ width: progress + '%' }}></div>
         </div>
     );
