+++ src/components/profile/ProfileHeader.js
@@ -9,29 +9,14 @@ import { useNavigate } from 'react-router-dom';
 //context
 import { UserContext } from '../../context/UserContext';
 //firebase
-import { getAuth, updateProfile, deleteUser, reauthenticateWithCredential, EmailAuthProvider } from 'firebase/auth';
-import { updateDoc, doc, onSnapshot } from 'firebase/firestore';
-import { getStorage, ref, uploadBytesResumable, getDownloadURL } from "firebase/storage";
-import { db } from '../../firebase.config';
-//toastify
-import { toast } from 'react-toastify';
-//components
-import ProgressBar from '../ProgressBar';
+import { getAuth, } from 'firebase/auth';
 
-const ProfileHeader = ({ setDisplayName, setPhoto }) => {
+const ProfileHeader = ({ displayName, setIsEdit, photo }) => {
 
     // ******** STATES AND OTHERS ********//
 
-    const { logout } = useLogout()
-    const { getData } = useExpensesIncomes()
     const { user, authIsReady } = useContext(UserContext)
     const auth = getAuth()
-    const navigate = useNavigate()
-    const [isEdit, setIsEdit] = useState(false)
-    const [loading, setLoading] = useState(false)
-    // const [photo, setPhoto] = useState(null)
-    // const [displayName, setDisplayName] = useState(null)
-    const [progressWidth, setProgressWidth] = useState(0)
     const [userData, setUserData] = useState({
         displayName: '',
         email: '',
@@ -39,7 +24,7 @@ const ProfileHeader = ({ setDisplayName, setPhoto }) => {
         expensesIncomes: []
     })
 
-    let { displayName, email, photoURL } = userData
+    let { email } = userData
 
     // ********* MAIN LOGIC *********//
 
@@ -59,183 +44,28 @@ const ProfileHeader = ({ setDisplayName, setPhoto }) => {
         setIsEdit(true)
     }
 
-    //deactivate edit mode: save edited info to firebase 
-    const handleTickClick = async () => {
-        const auth = getAuth()
-        //update details in firebase
-        try {
-            if (auth.currentUser.displayName !== displayName || auth.currentUser.photoURL !== photoURL) {
-                setLoading(true)
-                await updateProfile(auth.currentUser, {
-                    displayName,
-                    photoURL
-                })
-                //update details in firestore
-                const userRef = doc(db, 'users', auth.currentUser.uid)
-                await updateDoc(userRef, {
-                    displayName,
-                    photoURL
-                    // email
-                })
-                //add the storage photo url and display to state
-                setPhoto(photoURL)
-                setDisplayName(displayName)
-                //setLoading(false)
-                toast.success('Profile updated')
-            }
-        } catch (error) {
-            setLoading(false)
-            toast.error('Could not update profile details')
-            console.log(error);
-        }
-        setIsEdit(false)
-    }
-
-    //handle photo change
-    const handlePhotoChange = (e) => {
-        //check image size, max 2mb
-        if (e.target.files[0].size > 2000000) {
-            toast.error('Sorry, image size must be 2MB or lower')
-            e.target.value = ''
-        }
-        storeImage(e.target.files[0])
-
-    }
-
-    //save edited info in state
-    const handleEditChange = (e) => {
-        setUserData((prevState) => ({
-            ...prevState,
-            [e.target.id]: e.target.value
-        }))
-    }
-
-    //store image in firebase & get the download URL
-    //https://firebase.google.com/docs/storage/web/upload-files
-    const storeImage = async (image) => {
-        setLoading(true)
-        const auth = getAuth()
-        return new Promise((resolve, reject) => {
-            const storage = getStorage()
-            const fileName = `${ auth.currentUser.uid }-profile-picture`
-            const storageRef = ref(storage, 'images/' + fileName)
-            const uploadTask = uploadBytesResumable(storageRef, image);
-
-            uploadTask.on('state_changed',
-                (snapshot) => {
-                    // Get task progress, including the number of bytes uploaded and the total number of bytes to be uploaded
-                    const progress = (snapshot.bytesTransferred / snapshot.totalBytes) * 100;
-                    console.log(progress);
-                    setProgressWidth(progress)
-                    setTimeout(() => {
-                        setProgressWidth(0)
-                    }, 2000)
-                    switch (snapshot.state) {
-                        case 'paused':
-                            console.log('Upload is paused');
-                            break;
-                        case 'running':
-                            console.log('Upload is running');
-                            break;
-                    }
-                },
-                (error) => {
-                    reject(error)
-                },
-                () => {
-
-                    getDownloadURL(ref(storage, 'images/' + fileName))
-                        .then((url) => setUserData((prevState) => ({
-                            ...prevState,
-                            photoURL: url
-                        })))
-                }
-            );
-            setLoading(false)
-        })
-    }
-
-    //delete profile
-    const handleProfileDelete = async () => {
-        if (window.confirm('Are you sure you want to delete your profile?') === true) {
-
-            let email = window.prompt('Enter Email')
-            let password = window.prompt('Enter Password')
-
-            const auth = getAuth();
-            const user = auth.currentUser;
-            const credential = EmailAuthProvider.credential(email, password);
-
-            await reauthenticateWithCredential(user, credential).then(() => {
-                // User re-authenticated.
-                deleteUser(user).then(() => {
-                    getData()
-                    logout()
-                }).catch((error) => {
-                    toast.error('Could not delete user profile')
-                    console.log(error);
-                });
-            }).catch((error) => {
-                toast.error('Could not delete user profile')
-            });
-            navigate('/')
-        }
-    }
-
     return (
         <div className='profileHeader'>
-            <ProgressBar progress={progressWidth} />
-            {!isEdit && <Icon icon="mdi:cog" className='editProfileIcon'
+            <Icon icon="mdi:cog" className='editProfileIcon'
                 style={{ fontSize: '2rem' }}
-                onClick={handleCogClick} />}
-            {isEdit &&
-                <Icon icon="mdi:check-circle" className='editProfileIcon editProfileIconTick'
-                    style={{ fontSize: '2rem' }}
-                    onClick={handleTickClick} />
-            }
+                onClick={handleCogClick} />
 
-            {!isEdit ?
-                <div className="profilePictureContainer">
-                    <label htmlFor="photoURL">
-                        <img src={photoURL ? photoURL : require('../../assets/png/blank_profile.png')} alt="profile picture" />
-                    </label>
-                </div>
-                :
-                <div className="profilePictureContainer borderPulsate" >
-                    <label htmlFor="photoURL">
-                        <img src={photoURL ? photoURL : require('../../assets/png/blank_profile.png')} alt="profile picture" style={{ cursor: 'pointer' }} />
-                    </label>
-                    <input
-                        id="photoURL"
-                        type="file"
-                        max='1'
-                        accept='.jpg,.png,.jpeg'
-                        onChange={handlePhotoChange}
-                    />
-                </div>
-            }
+            <div className="profilePictureContainer">
+                <label htmlFor="photoURL">
+                    <img src={photo ? photo : require('../../assets/png/blank_profile.png')} alt="profile picture" />
+                </label>
+            </div>
 
-            {!isEdit
-                ?
-                <input
-                    type="text"
-                    className='profileName'
-                    id='displayName'
-                    minLength='3'
-                    maxLength='15'
-                    value={displayName}
-                    disabled />
-                :
-                <input
-                    type="text"
-                    className='profileName borderPulsate'
-                    id='displayName'
-                    minLength='3'
-                    maxLength='15'
-                    value={displayName}
-                    onChange={handleEditChange}
-                    style={{ border: '2px solid #2a9d8f' }}
-                />}
+
+            <input
+                type="text"
+                className='profileName'
+                id='displayName'
+                minLength='3'
+                maxLength='15'
+                value={displayName}
+                disabled />
+
 
             <input
                 type="text"
@@ -244,7 +74,6 @@ const ProfileHeader = ({ setDisplayName, setPhoto }) => {
                 value={email}
                 disabled />
 
-            <p className="deleteProfile" onClick={handleProfileDelete}>Delete Profile</p>
         </div>
     );
 }
